Don't move files, which don't exist

+++ makesymlinks.sh
@@ -25,7 +25,9 @@ echo "done"
 # move any existing dotfiles in homedir to dotfiles_old directory, then create symlinks from the homedir to any files in the ~/dotfiles directory specified in $files
 for file in $files; do
 	echo "Moving any existing dotfiles from ~ to $olddir"
-	mv ~/.$file ~/dotfiles_old/
+	if [ -f $file ]; then
+		mv ~/.$file ~/dotfiles_old/
+	fi
 	echo "Creating symlink to $file in home directory."
 	ln -s $dir/$file ~/.$file
 done
